% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{convergence.shapley}
\alias{convergence.shapley}
\title{Tests that the shapley algorithm converges.}
\usage{
convergence.shapley(row.nr, convergence.iterations = 30, iterations = 50,
  task = bh.task, model = train(makeLearner("regr.lm"), bh.task),
  return.value = "plot")
}
\arguments{
\item{row.nr}{Index for one observation of interest (not possible to choose a range of rows).
Input must be numeric.}

\item{convergence.iterations}{Amount of calls of the shapley function. Input must be one numeric.}

\item{iterations}{Amount of the iterations within the shapley function. Input must be one numeric.}

\item{return.value}{You can choose between plotting results or getting a data frame
with "plot" or "values".}
}
\value{
shapley values as a data.frame or a plot
}
\description{
Tests that the sum of values from the shapley algorithm converges against the
  difference of the mean of data and the prediction for the given observation. Supported tasks are
  classification, clustering and regression. Note that only for regression it is possible to choose
  to plot the result. For classification and clustering tasks, the predicted class (chosen with
  row.nr) is marked like this: >>ClassName<< . The result should show that this class was chosen most often.
}
