% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleyDataMean}
\alias{getShapleyDataMean}
\title{Getter function for a shapley object.}
\usage{
getShapleyDataMean(shapley.list)
}
\arguments{
\item{shapley.list}{A shapley object.}
}
\description{
Returns the mean of the target variable(s) from the given shapley object.
}
