% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleyFeatureNames}
\alias{getShapleyFeatureNames}
\title{Getter function for a shapley object.}
\usage{
getShapleyFeatureNames(shapley.list)
}
\arguments{
\item{shapley.list}{get shapley values}
}
\description{
Returns a vector of all feature names from the given shapley object.
}
