% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleyIds}
\alias{getShapleyIds}
\title{Getter function for a shapley object.}
\usage{
getShapleyIds(shapley.list)
}
\arguments{
\item{shapley.list}{A shapley object.}
}
\description{
Returns a vector of all unique `_Id`s from the given shapley object.
}
