% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleyPredictionResponse}
\alias{getShapleyPredictionResponse}
\title{Getter function for a shapley object.}
\usage{
getShapleyPredictionResponse(shapley.list)
}
\arguments{
\item{shapley.list}{A shapley object.}
}
\description{
Returns the prediction response from the given shapley object.
}
