% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleyPredictionType}
\alias{getShapleyPredictionType}
\title{Getter function for a shapley object.}
\usage{
getShapleyPredictionType(shapley.list)
}
\arguments{
\item{shapley.list}{A shapley object.}
}
\description{
Returns the prediction type from the given shapley object.
}
