% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleySubsetByResponseClass}
\alias{getShapleySubsetByResponseClass}
\title{Getter function for a shapley object.}
\usage{
getShapleySubsetByResponseClass(shapley.list)
}
\arguments{
\item{shapley.list}{A shapley object.}
}
\description{
This getter-function takes a shapley object. Unless the task type is "regr" it
  returns a shapley object with the same list elements except for the shapley.values. These are
  subseted to the response class.
}
