% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleyTaskType}
\alias{getShapleyTaskType}
\title{Getter function for a shapley object.}
\usage{
getShapleyTaskType(shapley.list)
}
\arguments{
\item{shapley.list}{get shapley values}
}
\description{
Returns the task type from the given shapley object.
}
