% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getShapleyValues}
\alias{getShapleyValues}
\title{Getter function for a shapley object.}
\usage{
getShapleyValues(shapley.list, subset = NULL, summate = FALSE)
}
\arguments{
\item{shapley.list}{A shapley object.}

\item{subset}{A vector that determines the rows.}

\item{summate}{Indicates whether the shapley values should be summed up or not. Default is 'FALSE'.}
}
\description{
Returns the shapley values from a given shapley object.
}
