% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapley.R
\name{plot.shapley.multipleFeatures}
\alias{plot.shapley.multipleFeatures}
\title{Plots a graph that shows the effect of several features over multiple observations.}
\usage{
\method{plot}{shapley.multipleFeatures}(shap.values, features = c("crim",
  "lstat"))
}
\arguments{
\item{shap.values}{A shapley object (generated by shapley(...)) that contains
the shapley.values and other important information about the task and model.}

\item{features}{A vector of the interesting feature names.}
}
\description{
This method draws a plot for, the observed value and describes
  the influence of the selected features.
}
