% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapley.R
\name{plot.shapley.singleValue}
\alias{plot.shapley.singleValue}
\title{Plots the difference for a single value.}
\usage{
\method{plot}{shapley.singleValue}(row.nr, shap.values = NULL,
  task = bh.task, model = train("regr.lm", bh.task))
}
\arguments{
\item{row.nr}{Index for the observation of interest.}

\item{shap.values}{(Optional) A data.frame that contains the shapley values. If
no shap.values are given the default algorithm is used to calculate the
shapley values for the given observation.}

\item{task}{mlr task that contains the data set.}

\item{model}{Model for the corresponding task.}
}
\description{
This method draws a plot for the data.mean, the observed value
  and describes the influence of features/variables for this difference.
}
