% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapleyr.R
\name{shapley}
\alias{shapley}
\title{Describes the difference between the expected value and the true outcome.}
\usage{
shapley(row.nr, task = bh.task, learner = "regr.lm", iterations = 50,
  method = "default")
}
\arguments{
\item{method}{Determines how the shapley value is calculated. Possible
selections are "default" or "kernel".}

\item{x}{A single observation of interest.}

\item{data}{The dataset that contains x.}

\item{algoirthm}{The algorithm that computes the expected value/class of x.}
}
\value{
shapley value as a matrix with col.names and their corresponding
  effects.
}
\description{
Takes an observation, the corresponding dataset where x is from
  and the algorithm that was applied to the dataset. With this input the
  shapley value for all features is calculated.
}
