% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_unsampl.R
\name{shapley.unsampled}
\alias{shapley.unsampled}
\title{Calculates the exact shapley value for every player based
on the algorithm for cooperative game theory.}
\usage{
shapley.unsampled(data.input = test, target = "value")
}
\arguments{
\item{target}{Put there the name of column with the results of the coalation-function.}
}
\description{
Calculates the exact shapley value for every player.
If you have N features/ players, there must be 2ˆN-1 rows in data.input. You can
leave the zero coalation away. Every row must be unique.
Below you can see an examle data input:
test.data = as.data.frame(rbind(c(1, 0, 5), c(0, 1, 4),c(1, 1, 10)))
name(test.data) = c("A", "B", "value")
And here with zero coalation:
test.data = as.data.frame(rbind(c(0, 0, 0), c(1, 0, 5), c(0, 1, 4),c(1, 1, 10)))
Player A has a coalation value of 5. Player A and B have a coalation value of 10.
}
