% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{test.convergence}
\alias{test.convergence}
\title{Tests that the shapley algorithm converges.}
\usage{
test.convergence(row.nr = 2, convergence.iterations = 20, iterations = 20,
  task = mtcars.task, model = train(makeLearner("cluster.kmeans"),
  mtcars.task), return.value = "values")
}
\arguments{
\item{row.nr}{Index for the observation of interest.}

\item{convergence.iterations}{Amount of calls of the shapley function.}

\item{iterations}{Amount of the iterations within the shapley function.}

\item{return.value}{You can choose between plotting results or getting a data frame}
}
\value{
shapley values as a data.frame or a plot
}
\description{
Tests that the shapley algorithm converges.
}
