% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{test.covergence}
\alias{test.covergence}
\title{Tests that the shapley algorithm converges.}
\usage{
test.covergence(row.nr, iterations = 20, shapley.iterations = 1)
}
\arguments{
\item{row.nr}{Index for the observation of interest.}

\item{iterations}{Amount of iterations.}

\item{shapley.iterations}{Amount of the iterations within the shapley
function.}
}
\value{
shapley value as a data.frame with col.names and their corresponding
  effects.
}
\description{
Tests that the shapley algorithm converges.
}
