% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapley.R
\name{compute.shapley.positions}
\alias{compute.shapley.positions}
\title{Calculates the positions of the features influence for plot.singleValue.}
\usage{
compute.shapley.positions(points, shift = 0)
}
\arguments{
\item{points}{A vector of shapley.values for a single row}

\item{shift}{data.mean}
}
\description{
Orders the values by their sign and value, shifts them and returns
  them as a vector.
}
