% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{convergence.shapley}
\alias{convergence.shapley}
\title{Tests that the shapley algorithm converges.}
\usage{
convergence.shapley(row.nr = 2, convergence.iterations = 100,
  iterations = 1, task = bh.task, model = train(makeLearner("regr.lm"),
  bh.task), return.value = "plot")
}
\arguments{
\item{row.nr}{Index for the observation of interest.}

\item{convergence.iterations}{Amount of calls of the shapley function.}

\item{iterations}{Amount of the iterations within the shapley function.}

\item{return.value}{You can choose between plotting results or getting a data frame}
}
\value{
shapley values as a data.frame or a plot
}
\description{
Tests that the sum of values from the shapley algorithm converges againts the
  difference of data.mean and the prediction for the given observation.
}
