% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapley.R
\name{plot.shapley.multipleValues}
\alias{plot.shapley.multipleValues}
\title{Plots a graph that shows the expected values, observed values and their difference.}
\usage{
\method{plot}{shapley.multipleValues}(shap.values)
}
\arguments{
\item{shap.values}{A shapley object (generated by shapley(...)) that contains
the shapley.values and other important information about the task and model.}
}
\description{
This method draws a plot that shows the mean of all observations, the observed
  values and the estimated values over multiple observcations.
}
