% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapleyR.R
\name{shapley}
\alias{shapley}
\title{Describes the difference between the expected value and the true outcome.}
\usage{
shapley(row.nr, task = bh.task, model = train(makeLearner("regr.lm"),
  bh.task), iterations = 50)
}
\arguments{
\item{row.nr}{Index for the observation of interest.}

\item{task}{mlr task that contains the data set.}

\item{model}{Model for the corresponding task..}

\item{iterations}{Amount of iterations.}
}
\value{
shapley value as a data.frame with col.names and their corresponding
  effects.
}
\description{
Takes an observation, the corresponding dataset where x is from
  and the algorithm that was applied to the dataset. With this input the
  shapley value for all features is calculated.
}
